Pull request fixes.
