+++ src/asgard_moment_manager.cpp
@@ -80,18 +80,17 @@ void moment_manager<P>::set_poisson(int const max_level, sparse_grid const &grid
                                     std::array<P, max_num_dimensions> const &xright,
                                     connection_patterns const &conn,
                                     hierarchy_manipulator<P> const &hier,
-                                    build_term_func<P> build_func,
-                                    iter_solve_func<P> iter_func)
+                                    build_term_func<P> build_func)
 {
   if (mlist.has_electric()) {
     moment_id const m0 = find_id(moment::zero(num_vel_));
     if (num_pos_ == 1) {
       moment_id const melectric = find_id(moment::electric(dimension_id(0), num_pos_));
       poisson_solver = poisson_1d<P>(hier.degree(), xleft[0], xright[0],
-                                              grid.current_level(0), m0, melectric);
+                                     grid.current_level(0), m0, melectric);
     } else {
       poisson_solver = poisson_md<P>(num_pos_, max_level, xleft, xright, conn, hier,
-                                              mlist, build_func, iter_func, m0);
+                                     mlist, build_func, m0);
     }
   }
 }
