+++ src/asgard_moment_manager.hpp
@@ -115,7 +115,13 @@ public:
 
   //! solves the poisson equation and caches them as electric moments
   void solve_poisson(sparse_grid const &grid, connection_patterns const &conn, hierarchy_manipulator<P> const &hier,
-                     interpolation_manager<P> const &interp, kronmult::workspace<P> &work, bool raw_on_cpu = false) const;
+                     interpolation_manager<P> const &interp, kronmult::workspace<P> &work) const {
+    constexpr bool raw_on_cpu = false;
+    solve_poisson(grid, conn, hier, interp, work, raw_on_cpu);
+  }
+  //! solves the poisson equation and caches them as electric moments
+  void solve_poisson(sparse_grid const &grid, connection_patterns const &conn, hierarchy_manipulator<P> const &hier,
+                     interpolation_manager<P> const &interp, kronmult::workspace<P> &work, bool raw_on_cpu) const;
   //! compute the specified interpolated moments
   void compute_interps(std::vector<moment_id> const &ids, sparse_grid const &grid,
                        std::vector<P> const &state, interpolation_manager<P> const &interp,
@@ -159,10 +165,24 @@ public:
   void compute_moments(std::vector<moment_id> const &mids, sparse_grid const &grid,
                        interpolation_manager<P> const &interp, connection_patterns const &conn,
                        hierarchy_manipulator<P> const &hier, kronmult::workspace<P> &kwork,
-                       gpu::vector<P> const &state, bool result_to_cpu = false) const;
-                   
+                       gpu::vector<P> const &state) const {
+    constexpr bool result_to_cpu = false;
+    compute_moments(mids, grid, interp, conn, hier, kwork, state, result_to_cpu);
+  }
+  //! load the given moments into the data-structures at device 0
+  void compute_moments(std::vector<moment_id> const &mids, sparse_grid const &grid,
+                       interpolation_manager<P> const &interp, connection_patterns const &conn,
+                       hierarchy_manipulator<P> const &hier, kronmult::workspace<P> &kwork,
+                       gpu::vector<P> const &state, bool result_to_cpu) const;
+  //! solves the poisson equation and caches them as electric moments, on gpu
   void solve_poisson_gpu(connection_patterns const &conn, hierarchy_manipulator<P> const &hier,
-                         interpolation_manager<P> const &interp, kronmult::workspace<P> &work, bool raw_on_cpu = false) const;
+                         interpolation_manager<P> const &interp, kronmult::workspace<P> &work) const {
+    constexpr bool raw_on_cpu = false;
+    solve_poisson_gpu(conn, hier, interp, work, raw_on_cpu);
+  }
+  //! solves the poisson equation and caches them as electric moments, on gpu
+  void solve_poisson_gpu(connection_patterns const &conn, hierarchy_manipulator<P> const &hier,
+                         interpolation_manager<P> const &interp, kronmult::workspace<P> &work, bool raw_on_cpu) const;
   #endif
   /*!
    * \brief Defines moments that should be used as raw or interpolation
@@ -183,8 +203,7 @@ public:
                    std::array<P, max_num_dimensions> const &xright,
                    connection_patterns const &conn,
                    hierarchy_manipulator<P> const &hier,
-                   build_term_func<P> build_func,
-                   iter_solve_func<P> iter_func);
+                   build_term_func<P> build_func);
 
   //! Updates the position grid if needed
   void update_position_grid(sparse_grid const &grid) const;
